requirements.txt
fastai
absl-py==0.12.0
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==5.0.7
defusedxml==0.7.1
entrypoints==0.3
gast==0.3.3
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
h5py==2.10.0
idna==2.10
ipykernel==5.5.3
ipython==7.23.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python-headless
opt-einsum==3.3.0
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.6.2
pyforest==1.0.3
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
pystan
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
Send2Trash==1.5.0
six==1.15.0
smmap==4.0.0
streamlit==0.81.0
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow-cpu
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.4
validators==0.18.2
watchdog==2.0.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
aiortc==1.2.0
av==8.0.3
streamlit-webrtc==0.11.0

